// src/model/Prefecture.ts
export interface Prefecture {
  projeto_id: number;
  data_prevista: string;
  reprogramacao: string;
  atraso: number;
  data_real_protocolo: string;
  sla_x: number;
  qtd_prefeitura: number;
  num_protocolo: string;
  status_operacional: string;
  data_aprov_prevista: string;
  data_aprov_real: string;
  sla_y: number;
  observacoes: string;
}